Only draw what's necessary
Redrawing the arcs leads to them having sharp edges as the smoothing
effect reduces with each redraw.  This version only draws the arcs that
differ from previously-drawn ones.  It does this by the following rule.

Given a number in 8-bit binary format, say 00011011, then the program
will only draw the arcs as specified by the leading 0001 part.  The
trailing 1011 is used to figure out where to start drawing (and at what
scale).  (Note that due to a previous commit, the instructions encoded
in 00011011 are read from right to left.)

So as we traverse a number, we use the 0s and 1s to keep track of where
we are until our remaining instructions look like 0...01.  At this point
we start drawing.  The exception is if we started with 0, but this is
easily dealt with.